refactor: layout 위치를 최상단으로 이동
